Stop the server gracefully when SIGTERM is sent (as docker stop does)
`docker stop` sends the `SIGTERM` signal to the process to stop it, but the
Minecraft server process handles this by quitting immediately, not saving the
world first. This differs from when the `stop` Minecraft command is used, as
that will save the world first before quitting.

This change adds a script (`start.sh`) which catches the `SIGTERM` signal and
handles it by sending the `stop` command to the Minecraft server. As a bonus,
`sendcommand` is now able to send commands to the server directly (by using a
named pipe) rather than using RCON.